Fix distance axis in sensitivity plots
As they were set, `min-dist` and `max-dist` were being ignored due
to a bug in pycbc_page_sensitivity (#685). After #685, the current
settings would not be suitable for O1 sensitivity plots. This patch
removes them, thus maintaining the previous behavior after #685 is
fixed. Additionally, it makes the summary range-vs-IFAR plot log in
distance and fixes the range to sensible values found in O1.