adjust splittable-num-banks to reduce job runtimes by a factor of 4 (#6)
* adjust splittable-num-banks to reduce job runtimes by a factor of 4

After observing in the OSG accounting data that we have lots of jobs running for >12 hours, beyond which we see a big uptick in badput (aka job re-runs due to preemption), this reduces the job runtime by a factor of 4 in order to reduce badput without making the average job runtime too short.

* correction to splittable-num-banks for workflow-splittable-full_data

I accidentally increased splittable-num-banks for workflow-splittable-full_data by only a factor of 2, not 4.  My intent (and the estimated runtimes I quote in my comment) was a factor of 4.