Removing gating from the test ini files
Removing the `[workflow-gating]` section from the test configuration. The test configuration, necessarily, must be appropriate for GWOSC data, and to my knowledge gating files have not been used there.

We don't need a special comment for how to include gating files, these are provided in the standard way for input files coming from outside the workflow now (which is to say that you just give the option in the section corresponding to the relevant executable along with the path, or URL, to the file, and the workflow code does the rest ... This is the same as how the PHASE_TIME statistic file is handled).